I added the file practica_juego_de_2.py
I added the file practica_juego_de_2.py, which is the second version of the code and the version with more funcinalidadesla first is that two players guess a secret number, the second is that there is a count of victories and the third function is that each player has a limit of errors.

if a player guesses the secret number he wins and that win is added to his win counter. if a player makes a mistake by giving the wrong number that is not the secret number, his win counter is decreased.