i added the file practica_juego_de_3.py
i added the file practica_juego_de_3.py, it is the third version of the code, this version has only two functionalities first is that two players guess a secret number, the second is that there is a win count. also the code is much more summarized.

if a player guesses the secret number he wins and that win is added to his win counter.